Rename idns into dnsi.
