+++ doc/dnsi.1
@@ -1,10 +1,10 @@
-.TH "idns" "1" "NLnet Labs"
+.TH "dnsi" "1" "NLnet Labs"
 
 .SH NAME
-idns - inspect the DNS
+dnsi - inspect the DNS
 
 .SH SYNOPSIS
-.B idns
+.B dnsi
 [<options>]
 <command>
 [<args>]
@@ -12,7 +12,7 @@ idns - inspect the DNS
 .SH DESCRIPTION
 Inspect the Domain Name System (DNS) in various ways.
 
-.B idns
+.B dnsi
 provides a number of commands that perform various tasks related to the DNS.
 
 Please consult the manual pages for these individual commands for more
@@ -21,6 +21,6 @@ information.
 .SH IDNS COMMANDS
 
 .PP
-\fBidns-query\fR(1)
+\fBdnsi-query\fR(1)
 .RS 4
 Send a query to a name server.
