+++ src/lib.rs
@@ -1,4 +1,4 @@
 //! The actual implementation of the tools.
 
-pub mod idns;
+pub mod dnsi;
